+++ dev-tool/Cargo.toml
@@ -13,6 +13,7 @@ displaydoc = "0.2.3"
 env_logger =  "0.9.0"
 indicatif = "0.16.2"
 itertools = "0.10.0"
+layer = { path = "../layer" }
 log = "0.4.14"
 ndarray = "0.15.3"
 rand = "0.8.3"
