+++ download_data.sh
@@ -46,4 +46,4 @@ download asset smbert v0000
 download asset qambert v0001
 download asset ltr v0000
 download misc bench_matmul v0000
-download misc ltr_test_data v0000
+download misc ltr_test_data v0001
