+++ layer/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "layer"
+version = "0.1.0"
+authors = ["Xayn Engineering <engineering@xaynet.dev>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bincode = "1.3.3"
+displaydoc = "0.2.3"
+# to be kept in sync with tract-core
+ndarray = "=0.15.3"
+rand = "0.8.4"
+rand_distr = "0.4.2"
+serde = { version = "1.0.130", features = ["derive"] }
+thiserror = "1.0.30"
+
+[dev-dependencies]
+test-utils = { path = "../test-utils" }
