+++ layer/src/lib.rs
@@ -1,0 +1,6 @@
+//! AI model building blocks.
+
+pub mod activation;
+pub mod dense;
+pub mod io;
+pub mod utils;
