+++ xayn-ai/Cargo.toml
@@ -10,12 +10,12 @@ bincode = "1.3.3"
 derive_more = { version = "0.99.16", default-features = false, features = ["from", "display"] }
 displaydoc = "0.2.3"
 itertools = "0.10.1"
+layer = { path = "../layer" }
 # to be kept in sync with rubert
 ndarray = "=0.15.3"
 # TODO: use version 1.0.5 once it is released
 obake = { git = "https://github.com/doctorn/obake", rev = "d6bea07e355ca4adf353c1e627f13c8c3286361b" }
 rand = "0.8.4"
-rand_distr = "0.4.2"
 rubert = { path = "../rubert" }
 serde = { version = "1.0.130", features = ["derive"] }
 serde_repr = "0.1.7"
